+++ vmngclient/models/configuration/feature_profile/sdwan/service/lan/svi.py
@@ -3,14 +3,14 @@ from typing import List, Optional, Union
 from pydantic import BaseModel, ConfigDict, Field
 
 from vmngclient.api.configuration_groups.parcel import Default, Global, Variable
-from vmngclient.model.configuration.feature_profile.sdwan.service.lan.common import (
+from vmngclient.models.configuration.common import RefId
+from vmngclient.models.configuration.feature_profile.sdwan.service.lan.common import (
     Arp,
     StaticIPv4Address,
     StaticIPv6Address,
     VrrpIPv6Address,
     VrrpTrackingObject,
 )
-from vmngclient.models.configuration.common import RefId
 
 
 class VrrpIPv4SecondaryAddress(BaseModel):
